Generate code using latest spec

+++ src/main/java/com/ringcentral/definitions/BillingInfo.java
@@ -5,18 +5,18 @@ public class BillingInfo {
     /**
      * Cost per minute, paid and already included in your RingCentral Plan. For example International Calls
      */
-    public Long costIncluded;
+    public Double costIncluded;
     /**
      * Cost per minute, paid and not included in your RingCentral Plan
      */
-    public Long costPurchased;
+    public Double costPurchased;
 
-    public BillingInfo costIncluded(Long costIncluded) {
+    public BillingInfo costIncluded(Double costIncluded) {
         this.costIncluded = costIncluded;
         return this;
     }
 
-    public BillingInfo costPurchased(Long costPurchased) {
+    public BillingInfo costPurchased(Double costPurchased) {
         this.costPurchased = costPurchased;
         return this;
     }
